src/Services/Detections/ai_inferencer/requirements.txt
dapr-ext-grpc
dapr
setuptools==58.2.0

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
# opencv-python-headless>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
psutil  # system utilization
avro-python3
avro-json-serializer
smmap
gitdb
gitpython
IPython
ultralytics
requests
boto3
onnxruntime